chore(travis): speed up dart2js unit tests
The dart2js jobs run times are dominated by slow Firefox runs.  In my
experiments, running multiple Firefox instances in parallel and sharding
the tests among them proved beneficial.

With this commit, the dart2js jobs take about 10m to run instead of 30m.
Compare these two recent Travis runs:
- 10m: <https://travis-ci.org/angular/angular.dart/builds/34415853>
  (this commit.)
- 30m: <https://travis-ci.org/angular/angular.dart/builds/34341163> (the
  commit just before this one.)

The following information is from earlier Travis runs.  Look at this
Travis run:
<https://travis-ci.org/angular/angular.dart/jobs/33159766>.  We can see
that the Firefox tests take 2m30s to run as opposed to the 25m [in a
regular run](https://travis-ci.org/angular/angular.dart/jobs/33152203).

There are actually 4 jobs running Firefox karma in parallel, each
running 1/4th of the normal tests.  The output is mingled but ok.

```text
Firefox 30.0.0 (Linux): Executed 409 of 409 SUCCESS (2 mins 16.855 secs / 2 mins 15.357 secs)
....................
..................................................
........................................
..................................................
Firefox 30.0.0 (Linux): Executed 410 of 410 SUCCESS (2 mins 14.153 secs / 2 mins 12.805 secs)
..................................................
........
Firefox 30.0.0 (Linux): Executed 408 of 408 SUCCESS (2 mins 34.253 secs / 2 mins 39.31 secs)
..................................................
...........
Firefox 30.0.0 (Linux): Executed 411 of 411 SUCCESS (2 mins 30.957 secs / 2 mins 29.93 secs)
```

This is how our traditional Firefox runs look:

```text
Firefox 30.0.0 (Linux): Executed 1565 of 1565 SUCCESS (25 mins 34.727 secs / 17 mins 13.582 secs)
```

Notes:
- The test output for the different shards are all intermingled.  This
  doesn't seem to be a big deal at the moment (they're all mostly just
  printing dots anyway) and we can certainly improve it later.  (GNU
  parallel instead of xargs would do this automatically but I think
  isn't installed and we don't have sudo.  We could roll our own.)
- Prereq: Commit 87279feb5096c3da29bc3014e8117d6d83cbf75a

+++ scripts/travis/install.sh
@@ -62,6 +62,15 @@ fi
 
 parallel_get() {(
   _download_and_unzip() {
+    # TODO(chirayu):  Hack to run npm in parallel.  Better to refactor and run
+    # both npm install and pub install in parallel.  Both must be run AFTER any
+    # rebasing onto g3v1x.  Also "pub install" can only happen after the Dart
+    # SDK is unzipped and the path has been updated.
+    if [[ "$1" == "npm" ]]; then
+      npm install
+      return
+    fi
+
     ZIPFILE=${1/*\//}
     curl -O -L $1 && unzip -q $ZIPFILE && rm $ZIPFILE
   }
@@ -70,7 +79,7 @@ parallel_get() {(
   echo "$@" | xargs -d ' ' -n 1 -P 2 -I URL bash -c '_download_and_unzip URL'
 )}
 
-parallel_get $DART_SDK_URL $DARTIUM_URL
+parallel_get npm $DART_SDK_URL $DARTIUM_URL
 
 echo Fetched new dart version $(<dart-sdk/version)
